### 2019-03-19 #1310 @gkthiruvathukal
@Gweissman and @mbod, Would either or both of you be willing to contribute a review to this JOSS submission? This is a Python NLP project and would closely match your areas of interest.
