### 2019-03-21 #1310 @gweissman
@gkthiruvathukal Thank you for thinking of me for the review. I could comment on things like documentation, ease of install, readability of software, and other user experience domains. However, I am completely unfamiliar with Uralic languages so could not comment on the accuracy of usefulness of the tool. Neither am I a computer scientist nor a linguist, so others may have more relevant expertise here. With the above caveats in mind, I'm happy to go ahead with a review if it would be helpful. Simultaneously, I'd be supportive of finding someone else with more relevant expertise who could offer substantive feedback.
