### 2019-03-21 #1310 @gkthiruvathukal
Thanks, @gweissman. I don't see too much expertise in Uralic languages on the board, so we're all kind of in the same boat. I am mainly interested in evaluating the aspects you mentioned. I'm hoping we will get one other reviewer (@mbod). I will assign you and wait a bit longer to see if we can get a second reviewer.  Then we'll start review.
