### 2019-03-22 #1345 @whedon
[REVIEW]: UralicNLP: An NLP library for Uralic Languages
**Submitting author:** @mikahama (<a href="http://orcid.org/0000-0001-9315-1278">Mika K Hämäläinen</a>)
**Repository:** <a href="https://github.com/mikahama/uralicNLP" target ="_blank">https://github.com/mikahama/uralicNLP</a>
**Version:** 1.1.1-joss

- [x] **Version:** 1.1.1-joss
