### 2019-04-02 #1345 @gweissman
@arfon This my first time reviewing for JOSS so still trying to learn the checklist, which I have completed above. Additional comments I would like to include in the review are:

- Most of the code examples in the README.md file work with the following exception: when I type `cg.disambiguate(tokens)` I get the following error:

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/local/lib/python3.7/site-packages/uralicNLP/cg3.py", line 39, in disambiguate
    cg_conv = Popen(["cg-conv" ,"-f"], stdout=PIPE, stdin=p1.stdout)
  File "/usr/local/Cellar/python/3.7.3/Frameworks/Python.framework/Versions/3.7/lib/python3.7/subprocess.py", line 775, in __init__
    restore_signals, start_new_session)
  File "/usr/local/Cellar/python/3.7.3/Frameworks/Python.framework/Versions/3.7/lib/python3.7/subprocess.py", line 1522, in _execute_child
    raise child_exception_type(errno_num, err_msg, err_filename)
FileNotFoundError: [Errno 2] No such file or directory: 'cg-conv': 'cg-conv'
```

- When I type `uralicApi.download("fin")` the download appears to work correctly, and a status message is completed at the end of each download. These took a while and I would suggest providing a status update prior to initiating the download rather than after it has completed so the user knows the program is not frozen.
