### 2019-04-02 #1345 @mikahama
@gweissman It seems that you do not have VISL CG-3 installed. It is required for constraint grammars and unfortunately cannot be installed over pip as a dependency. This requirement has been stated in the readme file.

I think that adding an indicator (or a progress bar) for downloading the models is a great idea. Thanks for the suggestion.
