### 2019-04-02 #1345 @gweissman
@mikahama You are correct this is mentioned in the README.md file. I have nothing further to add. Congratulations on what appears to be a very useful software package.
