### 2019-04-26 #1345 @gweissman
@gkthiruvathukal I have completed my review. Some of the check boxes I am unable to adequately assess. For example, I don't speak or read any of the languages in question so I can't make claims about functionality of the software.

- Not all of the references contained a DOI
- I did not see explicit community guidelines for participation
- I am unable to determine if there are automated tests in this software package
