+++ .coveralls.yml
@@ -1,0 +1,3 @@
+service_name: travis-ci
+repo_token: nzVGk44eE9ikkabuKey1zAbXY0com9o7R
+parallel: true
