+++ README.md
@@ -1,6 +1,6 @@
 # UralicNLP
 
-[![Build Status](https://travis-ci.com/mikahama/uralicNLP.svg?branch=master)](https://travis-ci.com/mikahama/uralicNLP)
+[![Build Status](https://travis-ci.com/mikahama/uralicNLP.svg?branch=master)](https://travis-ci.com/mikahama/uralicNLP) [![Updates](https://pyup.io/repos/github/mikahama/uralicNLP/shield.svg)](https://pyup.io/repos/github/mikahama/uralicNLP/)
 
 UralicNLP is a natural language processing library for small Uralic languages. Currently its functionality is provided by [akusanat.com](https://akusanat.com) API which is also developed by [Mika Hämäläinen](https://mikakalevi.com).
 
@@ -23,7 +23,21 @@ The API is under constant development and new languages will be added to the San
     from uralicNLP import uralicApi
     uralicApi.supported_languages()
     >>{'languages': ['sms', 'izh', 'mhr', 'vot', 'olo', 'myv', 'mdf', 'mrj', 'udm', 'yrk', 'koi', 'fin']}
-  
+
+### Download the models 
+
+If you have a lot of data to process, it might be a good idea to download the morphological models to your computer locally. This can be done easily.
+
+    from uralicNLP import uralicApi
+    uralicApi.download("fin")
+
+When models are installed, *generate()*, *analyze()* and *lemmatize()* methods will automatically use them instead of the server side API. [More information about the models](https://github.com/mikahama/uralicNLP/wiki/Models).
+
+Use **uralicApi.model_info(language)** to see information about the FSTs and CGs such as license and authors. If you know how to make this information more accurate, please don't hesitate to open an issue on GitHub
+
+    from uralicNLP import uralicApi
+    uralicApi.model_info("fin")
+
 ### Lemmatize words
 A word form can be lemmatized with UralicNLP. This does not do any disambiguation but rather returns a list of all the possible lemmas.
 
@@ -51,14 +65,16 @@ From a lemma and a morphological analysis, it's possible to generate the desired
   
 An example of generating the singular partitive form for the Finnish noun *käsi*. The result is *kättä*. The default generator is a **normative dictionary** generator. *uralicApi.generate("käsi+N+Sg+Par", "fin", dictionary_forms=False)* uses a regular normative generator and *uralicApi.generate("käsi+N+Sg+Par", "fin", descrpitive=True)* a descriptive generator.
 
-### Download the models to speed things up
 
-If you have a lot of data to process, it might be a good idea to download the morphological models to your computer locally. This can be done easily.
+### Access the HFST transducer
+
+If you need to get a lower level access to [the HFST transducer object](https://hfst.github.io/python/3.12.1/classhfst_1_1HfstTransducer.html), you can use th following code
 
     from uralicNLP import uralicApi
-    uralicApi.download("fin")
+    sms_generator = uralicApi.get_transducer("sms", analyzer=False) #generator
+    sms_analyzer = uralicApi.get_transducer("sms", analyzer=True) #analyzer
 
-When models are installed, *generate()*, *analyze()* and *lemmatize()* methods will automatically use them instead of the server side API. [More information about the models](https://github.com/mikahama/uralicNLP/wiki/Models).
+The same parameters can be used here as for *generate()* and *analyze()* to specify whether you want to use the normative or descriptive analyzers and so on. The defaults are *get_transducer(language, cache=True, analyzer=True, descrpitive=True, dictionary_forms=True)*.
 
 ### Syntax - Constraint Grammar disambiguation
 
[…]
