### 2019-04-30 e929203
Update CONTRIBUTING.md
