### 2019-05-02 #1345 @mikahama
As far as DOIs go, I have improved the situation with the latest version of UralicNLP. It is now possible to use uralicApi.model_info("fin") which will output information about the authors, copyright and such details. At least this will output a url to the Giellatekno repository where the FST and CG models are from. I have also gathered as much information as possible from the files in the Giellatekno repository in the model_info function. Sadly, it is quite common that the AUTHORS and LICENCE files are empty (see https://victorio.uit.no/langtech/trunk/langs/nob/ for example)

I have added CONTRIBUTING.md to the repository.

Also, I have improved the coverage of the automated tests (see https://github.com/mikahama/uralicNLP/blob/master/travis_test.py), however I do admit that more work should be done here to cover all the functionalities.
