### 2019-05-06 #1345 @gweissman
@gkthiruvathukal I have no further comments. @mikahama has responded adequately to the points raised in my review.
