### 2019-05-06 #1345 @gkthiruvathukal
Thanks, @gweissman.  And thanks, @mikahama for addressing all review feedback.

I have gone through the checklist, and I'm ready to accept.

@mikahama Can you please archive the accepted version of the package and provide an associated DOI (e.g., via Zenodo)?
