### 2019-05-06 #1345 @mikahama
It is now on Zenodo https://zenodo.org/record/2668061
