### 2019-05-09 #1345 @labarba
hi @mikahama — the JOSS paper format has the software repository on the left-margin of the first page, therefore we ask that you don't add the repository again in the text/footnote. Can you fix?
