### 2019-05-09 #1345 @labarba
Also, can you edit the Zenodo metadata so the title matches the JOSS paper? You could also add your ORCID.
