### 2019-05-09 #1345 @mikahama
@labarba Hi, I removed the Github link in the footnote and edited the Zenodo metadata.
