### 2019-05-09 #1345 @labarba
@mikahama — I'm going to ask for a little clean-up on your bibliography. When you have acronyms, like NLP, you should use caps protection only around that string, i.e., `{NLP}`, and not the whole title. For example, this line will override the bibliography style for title capitalization by caps protection on the whole title:
https://github.com/mikahama/uralicNLP/blob/677a54022a21abb1d863de0b78f7bc8514b95232/paper/paper.bib#L3

For two entries, I found you can add a DOI or permalink:

- `yokohama`  add permalink http://hdl.handle.net/10138/282733
- `linden2013hfst` add DOI https://doi.org/10.1007/978-3-642-40486-3_4
