+++ paper/paper.bib
@@ -1,26 +1,27 @@
 @inproceedings{euraleksi,
 author = {H{\"a}m{\"a}l{\"a}inen, Mika and Rueter, Jack},
-title = {{Advances in Synchronized XML-MediaWiki Dictionary Development in the Context of Endangered Uralic Languages}},
-booktitle = {{Proceedings of the Eighteenth EURALEX International Congress}},
+title = {Advances in Synchronized {XML-MediaWiki} Dictionary Development in the Context of Endangered {U}ralic Languages},
+booktitle = {Proceedings of the Eighteenth {EURALEX} International Congress},
 pages = {967-978},
 year = {2018}
 
 },
 @inbook{BeesleyKarttunen03,
   author = {Kenneth R. Beesley and Lauri Karttunen},
-  title = {{Finite-State Morphology}},
-  publisher = {{Stanford, CA: CSLI Publications}},
+  title = {Finite-State Morphology},
+  publisher = {Stanford, {CA}: {CSLI} Publications},
   pages={451--454},
   year = 2003
 },
  @InProceedings{yokohama,
   author = {Mika H{\"a}m{\"a}l{\"a}inen},
-  title = "{Extracting a Semantic Database with Syntactic Relations for Finnish to Boost Resources for Endangered Uralic Languages}",
-  booktitle = {Proceedings of the Logic and Engineering of Natural Language Semantics 15 (LENLS15)},
-  year = {2018}
+  title = {Extracting a Semantic Database with Syntactic Relations for {F}innish to Boost Resources for Endangered {U}ralic Languages},
+  booktitle = {Proceedings of the Logic and Engineering of Natural Language Semantics 15 ({LENLS15})},
+  year = {2018},
+  url= {https://helda.helsinki.fi/handle/10138/282733}
   },
  @inproceedings{poem_gen,
-  title={{Harnessing NLG to Create Finnish Poetry Automatically}},
+  title={Harnessing {NLG} to Create {F}innish Poetry Automatically},
   author={H{\"a}m{\"a}l{\"a}inen, Mika},
   booktitle={Proceedings of the Ninth International Conference on Computational Creativity},
   pages={{9-15}},
@@ -30,11 +31,11 @@ year = {2018}
 author = {Bick, Eckhard and Didriksen, Tino},
 institution = {University of Southern Denmark, Denmark},
 institution = {GrammarSoft ApS, Denmark},
-booktitle = {Proceedings of the 20th Nordic Conference of Computational Linguistics, NODALIDA 2015, May 11-13, 2015, Vilnius, Lithuania},
+booktitle = {Proceedings of the 20th Nordic Conference of Computational Linguistics, {NODALIDA} 2015, May 11-13, 2015, Vilnius, Lithuania},
 number = {109},
 pages = {31-39},
 publisher = {Linköping University Electronic Press, Linköpings universitet},
-title = {CG-3 — Beyond Classical Constraint Grammar},
+title = {{CG-3} — Beyond Classical Constraint Grammar},
 year = {2015},
 issn = {1650-3740},
 keywords = {},
@@ -45,7 +46,7 @@ author = {Moshagen, Sjur N. and Pirinen, Tommi A. and Trosterud, Trond},
 institution = {University of Tromsø, Norway},
 institution = {Helsinki university, Finland},
 institution = {University of Tromsø, Norway},
-booktitle = {Proceedings of the 19th Nordic Conference of Computational Linguistics (NODALIDA 2013); May 22-24; 2013; Oslo University; Norway. NEALT Proceedings Series 16},
+booktitle = {Proceedings of the 19th Nordic Conference of Computational Linguistics ({NODALIDA} 2013); May 22-24; 2013; Oslo University; Norway. NEALT Proceedings Series 16},
[…]
