### 2019-05-09 #1345 @mikahama
@labarba I fixed the bibliography
