### 2019-05-09 #1345 @labarba
Congratulations, @mikahama, your JOSS paper is published! 🎉

Huge thanks to our editor: @gkthiruvathukal, and reviewer: @gweissman — we could not do this without you 🙏
